requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyascii==0.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
blis==1.1.0
bs4==0.0.2
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
defusedxml==0.7.1
Django==5.1.4
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
Flask==3.1.0
h11==0.16.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
l18n==2021.3
laces==0.1.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
mypy-extensions==1.0.0
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pathspec==0.12.1
pefile==2023.2.7
pillow==10.4.0
pillow_heif==0.21.0
platformdirs==4.3.6
preshed==3.0.9
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32-ctypes==0.2.2
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.31.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.5.0
telepath==0.3.1
thinc==8.3.3
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3<2.0.0
wagtail==6.3.1
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
Willow==1.9.0
wrapt==1.17.0
wsproto==1.2.0
